feat: new integration hooks
