+++ src/services/user/userIntegrationsApi/index.ts
@@ -2,11 +2,11 @@ import { AxiosResponse } from "axios";
 import Integration from "types/entities/Integration";
 import { apiGet, apiPost } from "..";
 
-const integrationsApi = {
+const userIntegrationsApi = {
   getIntegration: (): Promise<AxiosResponse<Integration>> => apiGet(`integration`),
 
   createIntegration: (data: any): Promise<AxiosResponse<Integration>> =>
     apiPost("integration", data),
 };
 
-export default integrationsApi;
+export default userIntegrationsApi;
